chore: wrap stories in lazyloader
