+++ botfront/imports/ui/components/stories/StoriesPageMenu.jsx
@@ -7,17 +7,22 @@ import {
 import React from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
+import { withTracker } from 'meteor/react-meteor-data';
+import { Instances } from '../../../api/instances/instances.collection';
+import { Projects } from '../../../api/project/project.collection';
 import { isTraining } from '../../../api/nlu_model/nlu_model.utils';
 import TrainButton from '../utils/TrainButton';
 import { PageMenu } from '../utils/Utils';
 
-export default function StoriesPageMenu(props) {
+function StoriesPageMenu(props) {
     const {
         project,
         project: { _id: projectId, training: { endTime, status } = {} },
         instance,
+        ready,
     } = props;
 
+    if (!ready) return null;
     return (
         <PageMenu title='Stories' icon='book'>
             <Menu.Menu position='right'>
@@ -87,6 +92,34 @@ export default function StoriesPageMenu(props) {
 }
 
 StoriesPageMenu.propTypes = {
+    ready: PropTypes.bool.isRequired,
     project: PropTypes.any.isRequired,
-    instance: PropTypes.object.isRequired,
+    instance: PropTypes.object,
 };
+
+StoriesPageMenu.defaultProps = {
+    instance: null,
+};
+
+export default withTracker((props) => {
+    const { projectId } = props;
+    const projectsHandler = Meteor.subscribe('projects', projectId);
+    const instancesHandler = Meteor.subscribe('nlu_instances', projectId);
+    const { training } = Projects.findOne(
+        { _id: projectId },
+        {
+            fields: {
+                training: 1,
+            },
+        },
+    );
+    const instance = Instances.findOne({ projectId });
+
+    return {
+        ready:
+            projectsHandler.ready()
+            && instancesHandler.ready(),
+        instance,
+        project: { _id: projectId, training },
+    };
+})(StoriesPageMenu);
